refactor(headless): dedupe shared CLI/harness/seed helpers; drop dead exports
Extract parseArgs + chromeLaunchArgs into headless/lib/cli.mjs (the
hardcoded GPU_ARGS copies in probe/smoke/media-smoke were Windows-only;
they now get the platform-correct ANGLE backend + env override). edit.mjs
reuses render-core's startHarness (generalized for the media-less edit
path) instead of its own copy. Share collectAddClipMedia (workspace.mjs)
and seedMediaLibrary (src/headless/seed-media.ts). Remove unused
deleteRenderQueue + exportsDir exports; simplify abortJob to void.
